rename _py_op_commands to _op_commands
