+++ ipython_snippets/pytest/misc.py
@@ -2,7 +2,7 @@ import os
 
 from pyonepassword import OP, logging
 from pyonepassword._op_cli_config import OPCLIConfig
-from pyonepassword._py_op_commands import EXISTING_AUTH_REQD
+from pyonepassword._op_commands import EXISTING_AUTH_REQD
 from tests.fixtures.op_fixtures import _setup_alt_env, _setup_normal_env
 from tests.fixtures.valid_op_cli_config import (
     VALID_OP_CONFIG_NO_SHORTHAND_KEY,
