+++ pyonepassword/api/authentication.py
@@ -1,4 +1,4 @@
-from .._py_op_commands import (
+from .._op_commands import (
     EXISTING_AUTH_AVAIL,
     EXISTING_AUTH_IGNORE,
     EXISTING_AUTH_REQD,
